requirements.txt
pydantic==1.9.1
fastapi==0.78.0
uvicorn[standard]===0.17.6
pymongo==4.1.1
motor==3.1.1
python-decouple==3.6